+++ src/be/vdab/entities/Genre.java
@@ -2,17 +2,17 @@ package be.vdab.entities;
 
 public class Genre {
 	
-	private int id;
+	private long id;
 	private String naam;
 		
-	public Genre(int id, String naam) {
+	public Genre(long id, String naam) {
 		this.id = id;
 		this.naam = naam;
 	}
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getNaam() {
